refactor(model_estimation): transform _fit_transfer_ into _model_estimation_
